requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.10.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens @ file:///C:/b/abs_9662ywy9fp/croot/asttokens_1743630464377/work
async-lru==2.0.5
attrs==25.3.0
av==15.1.0
babel==2.17.0
beautifulsoup4==4.13.5
bleach==6.2.0
Brotli==1.1.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/colorama_1699472650914/work
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.17
decorator @ file:///C:/miniconda3/conda-bld/decorator_1757341251593/work
defusedxml==0.7.1
einops==0.8.1
executing @ file:///C:/miniconda3/conda-bld/executing_1757061247746/work
fastapi==0.116.2
fastjsonschema==2.21.2
ffmpy==0.6.1
filelock==3.19.1
fonttools==4.60.0
fqdn==1.5.1
fsspec==2025.9.0
ftfy==6.3.1
fvcore==0.1.5.post20221221
gradio==5.46.1
gradio_client==1.13.1
groovy==0.1.2
grpcio==1.75.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-ws==0.7.2
huggingface==0.0.1
huggingface-hub==0.35.0
hyperframe==6.1.0
idna==3.10
iopath==0.1.10
ipykernel==6.30.1
ipython @ file:///C:/b/abs_307wfft5yu/croot/ipython_1744912982439/work
ipython_pygments_lexers @ file:///C:/b/abs_b66hj1lo19/croot/ipython_pygments_lexers_1744753262043/work
isoduration==20.11.0
jedi @ file:///C:/b/abs_3a2kbnlclc/croot/jedi_1733987412687/work
Jinja2==3.1.6
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.7
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.9
lark==1.2.2
lightning-utilities==0.15.2
lmstudio==1.5.0
loguru==0.7.3
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
matplotlib-inline @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/matplotlib-inline_1699484796387/work
mdurl==0.1.2
mistune==3.1.4
mpmath==1.3.0
msgspec==0.19.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
notebook==7.4.5
notebook_shim==0.2.4
numpy==2.3.3
orjson==3.11.3
packaging==25.0
pandas==2.3.2
pandocfilters==1.5.1
parameterized==0.9.0
parso @ file:///C:/b/abs_834b4mj92b/croot/parso_1733963322289/work
pillow==11.3.0
platformdirs==4.4.0
portalocker==3.2.0
prometheus_client==0.23.1
prompt-toolkit @ file:///C:/b/abs_68uwr58ed1/croot/prompt-toolkit_1704404394082/work
protobuf==6.32.1
psutil==7.1.0
pure_eval @ file:///C:/miniconda3/conda-bld/pure_eval_1757067062212/work
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
pydub==0.25.1
Pygments @ file:///C:/b/abs_e4bg5vh5j_/croot/pygments_1744667628203/work
pyparsing==3.2.4
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
python-multipart==0.0.20
pytorchvideo @ git+https://github.com/facebookresearch/pytorchvideo.git@6cdc929315aab1b5674b6dcf73b16ec99147735f
pytz==2025.2
pywin32==311
pywinpty==3.0.0
PyYAML==6.0.2
pyzmq==27.1.0
qdrant-client==1.15.1
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.1.0
rpds-py==0.27.1
ruff==0.13.1
safehttpx==0.1.6
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==1.8.3
setuptools==78.1.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
stack_data @ file:///C:/miniconda3/conda-bld/stack_data_1757067057854/work
starlette==0.48.0
sympy==1.14.0
tabulate==0.9.0
termcolor==3.1.0
terminado==0.18.1
threadpoolctl==3.6.0
timm==1.0.19
tinycss2==1.4.0
tokenizers==0.22.1
tomlkit==0.13.3
torch==2.8.0+cu126
torchaudio==2.8.0
torchmetrics==1.8.1
torchvision==0.23.0+cu126
tornado==6.5.2
tqdm==4.67.1
traitlets @ file:///C:/b/abs_bfsnoxl4pq/croot/traitlets_1718227069245/work
transformers==4.56.2
typer==0.18.0
types-python-dateutil==2.9.0.20250822
types-regex==2025.9.1.20250903
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.35.0
wcwidth @ file:///C:/b/abs_0c8p4c33bp/croot/wcwidth_1750352902378/work
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
wheel==0.45.1
win32_setctime==1.2.0
wsproto==1.2.0
yacs==0.1.8
